requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
streamlit>=1.44.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
sqlalchemy>=2.0.32
httpx>=0.27.0
tenacity>=8.5.0
python-dotenv>=1.0.1
openai>=1.40.0
google-cloud-aiplatform>=1.70.0
pytest>=8.3.0
